Staging

+++ src/entities/deposits.rs
@@ -1,9 +1,10 @@
 use chrono::{DateTime, Utc};
+use bigdecimal::BigDecimal;
 
 table! {
     user_deposits (user_acct) {
         user_acct -> Varchar,
-        token_amount -> Int8,
+        token_amount -> Numeric,
         mint_acct -> Varchar,
         tx_sig -> Varchar,
         created_at -> Timestamptz,
@@ -15,14 +16,14 @@ table! {
 #[diesel(check_for_backend(diesel::pg::Pg))]
 pub struct UserDeposit {
     pub user_acct: String,
-    pub token_amount: i64,
+    pub token_amount: BigDecimal,
     pub mint_acct: String,
     pub tx_sig: String,
     pub created_at: DateTime<Utc>,
 }
 
 impl UserDeposit {
-    pub fn new(user_acct: String, token_amount: i64, mint_acct: String, tx_sig: String) -> Self {
+    pub fn new(user_acct: String, token_amount: BigDecimal, mint_acct: String, tx_sig: String) -> Self {
         UserDeposit {
             user_acct,
             token_amount,
